+++ src/ConstantExpression.cpp
@@ -4,7 +4,7 @@ std::shared_ptr<Expression> ConstantExpression::diff(std::string &s) {
     return std::make_shared<ConstantExpression>(ConstantExpression(0));
 }
 
-double ConstantExpression::eval(std::map<std::string, int> &params) {
+double ConstantExpression::eval(std::map<std::string, double> &params) {
     return value;
 }
